refactor: remove defaults file
